Change docs to use Python module invocation
That is, `python -m easyucs.easyucs`-style invocation. This updates the
references in the documentation to commands formerly `easyucs.py` and
`easyucs_gui.py`. It also changes install instructions, and recommends
Pip as the primary installation method.

References to files that have moved are also adjusted.

Additionally, trailing spaces are removed.