Fix package_data not being recusively included
This resolves a mis-use of setuptools causing package_data
(specifically, `schema/`) not to be included recursively. Several
directories were previously specified to setuptools using globs like
`schema/**/*`. As it turns out, such globs do not recursively expand to
list all files unless an additional option is passed to the underlying
function, which setuptools does not seem to give an option to provide.

This commit adds a function to work around that behavior.