Fix error preventing logfile from being generated
This commit resolves issue #7 in which when `--logfile` is specified on
the command line, it is created but not written to. This was caused by a
line in the creation of the console logger dropping the file logger from
the list of linked log handlers.

Now when `--logfile` is passed, the file is created, and both it and
stderr are written to.